requirements.txt
pywin32
smolagents[toolkit]
smolagents[openai]
huggingface_hub
win32com
python-dotenv
opencv-python
numpy
# Phoenix tracing dependencies
arize-phoenix-otel
openinference-instrumentation-openai
openinference-semantic-conventions
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http